Sized byte-literals are zero-padded
Using the `IsString` instance of `BytesN n` can cause an error if the
provided string literal is too short, or too long. Especially for
string-literals such as `"ETH" :: BytesN 8`, this behaviour is not very
user friendly.

With this change too short literals are zero padded to the right.
If the literal is too long, the previous behaviour is preserved, and
an error is thrown.