Return `Maybe Block` from `getBlockBy*` methods
Problem: Infura docs¹ say that `getBlockByNumber` and its friend
return "A block object, or null when no block was found".
`getBlockByHash` and `getBlockByNumber` do not account for that
and throw `ParsingException` with auto-generated message when null
is received. That's quite inconvenient to handle because you need
to depend on this auto-generated message.

Solution: return `Maybe Block` instead of just `Block`.
`instance FromJSON (Maybe a)` is defined in such a way that
null is parsed as `Nothing` and everything else is parsed using `a`
parser, which is exactly what we want.
Note that `getTransactionByHash` for example already returns
`Maybe Transaction` and was inconsistent with `getBlockByHash`.

¹ https://infura.io/docs/ethereum/json-rpc/eth-getBlockByHash