getBlock* related improvements (#111)
* Return `Maybe Block` from `getBlockBy*` methods

Problem: Infura docs¹ say that `getBlockByNumber` and its friend
return "A block object, or null when no block was found".
`getBlockByHash` and `getBlockByNumber` do not account for that
and throw `ParsingException` with auto-generated message when null
is received. That's quite inconvenient to handle because you need
to depend on this auto-generated message.

Solution: return `Maybe Block` instead of just `Block`.
`instance FromJSON (Maybe a)` is defined in such a way that
null is parsed as `Nothing` and everything else is parsed using `a`
parser, which is exactly what we want.
Note that `getTransactionByHash` for example already returns
`Maybe Transaction` and was inconsistent with `getBlockByHash`.

¹ https://infura.io/docs/ethereum/json-rpc/eth-getBlockByHash

* Rename blockReceiptsRoot

Problem: in Infura docs¹ there is a field called `receiptsRoot`,
but in code it's called `blockReceiptRoot`, so apparently it's
always parsed as `Nothing` because parser is auto-generated from
the field name.

Solution: rename the field.

¹ https://infura.io/docs/ethereum/json-rpc/eth-getBlockByHash

* Add methods to download blocks without tx bodies

Problem: `eth_getBlockByHash` accepts a boolean argument that
determines whether full transaction objects will be included in
the response or only transaction hashes.
The existing API hardcodes this argument to True, but sometimes
passing False is desired (at least for performance reasons).

Solution:
1. Generalize `Block` to have a type parameter. Alternatively we
could define another type very similar to the existing `Block`, but
it would be some duplication of code.
2. Add `getBlockByHashLite` and `getBlockByNumberLite` methods that
pass False and return `BlockT HexString`.