Fix ABI encoding for fixed length lists
The ABI encoding for fixed length Solidity lists `T[k]` was wrong in the case that `T` is dynamic. If T is dynamic, then according to the [spec](https://docs.soliditylang.org/en/v0.5.3/abi-spec.html#formal-specification-of-the-encoding),

> The following types are called “dynamic”:
> `T[k]` for any dynamic `T` and any `k >= 0`

This PR ensures this is the case and handles encoding for dynamic `T` similar to how it is handled in the variable length array case for `T[]` which was fixed in #126.